# make two tests faster