--- a/tensorflow_addons/optimizers/stochastic_weight_averaging_test.py
+++ b/tensorflow_addons/optimizers/stochastic_weight_averaging_test.py
@@ -83,7 +83,7 @@ class SWATest(tf.test.TestCase):
             "sgd", start_averaging=num_examples // 32 - 1, average_period=100
         )
         model.compile(optimizer, loss="mse")
-        model.fit(x, y, epochs=10)
+        model.fit(x, y, epochs=2)
         optimizer.assign_average_vars(model.variables)
 
         x = np.random.standard_normal((100, 3))