stochastic_weight_averaging_test: 57s -> 15s
novograd_test  : 58s -> 14s

related to https://github.com/tensorflow/addons/issues/1143